Hackathon
These codes contain the basic design for the website that we're creating.  The website is made so that the farmers in India can easily sell their products. This aims to reduce the wastage of surplus products. Every year tonnes of argricultural yield is wasted. This can be reduced by using technologies such as ours.